+++ static/althea/src/ErrorBoundary.js
@@ -18,7 +18,6 @@ class ErrorBoundary extends React.Component {
 
   componentDidCatch(error, errorInfo) {
     this.setState({ error, errorInfo });
-    console.log(error, errorInfo);
   }
 
   render() {