+++ static/althea/src/Settings/Firmware/UpdateFirmware.js
@@ -29,7 +29,6 @@ export default ({ open, setOpen }) => {
     try {
       await post("/router/update");
     } catch (e) {
-      console.log(e);
     }
   };